+++ pkg/recollection/repository/pgx/fe_updates_get.sql
@@ -26,4 +26,4 @@ ORDER BY CASE
     CASE
         WHEN LENGTH(@search_query::text) = 0 THEN animelayer_updates.update_date 
     END DESC
-LIMIT @count OFFSET @offset_count;
+LIMIT @count::bigint OFFSET @offset_count::bigint;
